chore: organize files into folders
