+++ app/components/UI/Bridge/components/BridgeDestNetworkSelector/BridgeDestNetworkSelector.test.tsx
@@ -1,9 +1,9 @@
 import { fireEvent } from '@testing-library/react-native';
-import { renderScreen } from '../../../util/test/renderWithProvider';
-import { BridgeDestNetworkSelector } from './BridgeDestNetworkSelector';
-import Routes from '../../../constants/navigation/Routes';
+import { renderScreen } from '../../../../../util/test/renderWithProvider';
+import { BridgeDestNetworkSelector } from '.';
+import Routes from '../../../../../constants/navigation/Routes';
 import { Hex } from '@metamask/utils';
-import { setSelectedDestChainId } from '../../../core/redux/slices/bridge';
+import { setSelectedDestChainId } from '../../../../../core/redux/slices/bridge';
 import { BridgeFeatureFlagsKey } from '@metamask/bridge-controller';
 
 const mockNavigate = jest.fn();
