+++ app/components/UI/Bridge/components/BridgeSourceTokenSelector/BridgeSourceTokenSelector.test.tsx
@@ -1,9 +1,9 @@
 import { fireEvent, waitFor } from '@testing-library/react-native';
-import { renderScreen } from '../../../util/test/renderWithProvider';
-import { BridgeDestTokenSelector } from './BridgeDestTokenSelector';
-import Routes from '../../../constants/navigation/Routes';
+import { renderScreen } from '../../../../../util/test/renderWithProvider';
+import { BridgeSourceTokenSelector } from '.';
+import Routes from '../../../../../constants/navigation/Routes';
 import { Hex } from '@metamask/utils';
-import { setDestToken } from '../../../core/redux/slices/bridge';
+import { setSourceToken } from '../../../../../core/redux/slices/bridge';
 import { BridgeFeatureFlagsKey } from '@metamask/bridge-controller';
 
 const mockNavigate = jest.fn();
@@ -23,14 +23,13 @@ jest.mock('../../../core/redux/slices/bridge', () => {
     __esModule: true,
     ...actual,
     default: actual.default,
-    setDestToken: jest.fn(actual.setDestToken),
+    setSourceToken: jest.fn(actual.setSourceToken),
   };
 });
 
-describe('BridgeDestTokenSelector', () => {
+describe('BridgeSourceTokenSelector', () => {
   const mockAddress = '0x1234567890123456789012345678901234567890' as Hex;
-  const mockSourceChainId = '0x1' as Hex;
-  const mockDestChainId = '0xa' as Hex;
+  const mockChainId = '0x1' as Hex;
   const token1Address = '0x0000000000000000000000000000000000000001' as Hex;
   const token2Address = '0x0000000000000000000000000000000000000002' as Hex;
 
@@ -50,18 +49,19 @@ describe('BridgeDestTokenSelector', () => {
         TokenBalancesController: {
           tokenBalances: {
             [mockAddress]: {
-              [mockSourceChainId]: {
+              [mockChainId]: {
                 [token1Address]: '0x0de0b6b3a7640000' as Hex, // 1 TOKEN1
+                [token2Address]: '0x1bc16d674ec80000' as Hex, // 2 HELLO
               },
-              [mockDestChainId]: {
-                [token2Address]: '0x1bc16d674ec80000' as Hex, // 2 HELLO on Optimism
+              '0xa': {
+                [token1Address]: '0x4563918244f40000' as Hex, // 5 TOKEN1 on Optimism
               },
             },
           },
         },
         TokensController: {
           allTokens: {
-            [mockSourceChainId]: {
+            [mockChainId]: {
               [mockAddress]: [
                 {
                   address: token1Address,
@@ -71,10 +71,6 @@ describe('BridgeDestTokenSelector', () => {
                   name: 'Token One',
                   aggregators: ['1inch'],
                 },
-              ],
-            },
-            [mockDestChainId]: {
-              [mockAddress]: [
                 {
                   address: token2Address,
                   symbol: 'HELLO',
@@ -120,8 +116,8 @@ describe('BridgeDestTokenSelector', () => {
             },
           },
           networkConfigurationsByChainId:  {
-            [mockSourceChainId]: {
-              chainId: mockSourceChainId,
+            '0x1': {
+              chainId: '0x1' as Hex,
               rpcEndpoints: [
                 {
                   networkClientId: 'selectedNetworkClientId',
@@ -129,10 +125,9 @@ describe('BridgeDestTokenSelector', () => {
               ],
               defaultRpcEndpointIndex: 0,
               nativeCurrency: 'ETH',
-              name: 'Ethereum',
             },
-            [mockDestChainId]: {
-              chainId: mockDestChainId,
+            '0xa': {
+              chainId: '0xa' as Hex,
               rpcEndpoints: [
                 {
                   networkClientId: 'optimismNetworkClientId',
@@ -140,11 +135,10 @@ describe('BridgeDestTokenSelector', () => {
               ],
               defaultRpcEndpointIndex: 0,
               nativeCurrency: 'ETH',
-              name: 'Optimism',
             },
           },
           providerConfig: {
-            chainId: mockSourceChainId,
+            chainId: mockChainId,
             ticker: 'ETH',
             rpcPrefs: { blockExplorerUrl: 'https://etherscan.io' },
             type: 'infura',
@@ -157,12 +151,12 @@ describe('BridgeDestTokenSelector', () => {
             },
           },
           accountsByChainId: {
-            [mockSourceChainId]: {
+            [mockChainId]: {
               [mockAddress]: {
                 balance: '0x29a2241af62c0000' as Hex, // 3 ETH
               },
             },
-            [mockDestChainId]: {
+            '0xa': {
               [mockAddress]: {
                 balance: '0x1158e460913d00000' as Hex, // 20 ETH on Optimism
               },
@@ -197,18 +191,23 @@ describe('BridgeDestTokenSelector', () => {
         },
         TokenRatesController: {
           marketData: {
-            [mockSourceChainId]: {
+            [mockChainId]: {
               [token1Address]: {
                 tokenAddress: token1Address,
                 currency: 'ETH',
                 price: 10, // 1 TOKEN1 = 10 ETH
               },
-            },
-            [mockDestChainId]: {
               [token2Address]: {
                 tokenAddress: token2Address,
                 currency: 'ETH',
-                price: 5, // 1 TOKEN2 = 5 ETH on Optimism
+                price: 50, // 1 TOKEN2 = 5 ETH
+              },
+            },
+            '0xa': {
+              [token1Address]: {
+                tokenAddress: token1Address,
+                currency: 'ETH',
+                price: 8, // 1 TOKEN1 = 8 ETH on Optimism
               },
             },
           },
@@ -239,9 +238,18 @@ describe('BridgeDestTokenSelector', () => {
             },
           },
           tokensChainsCache: {
-            [mockDestChainId]: {
+            [mockChainId]: {
               timestamp: Date.now(),
               data: {
+                [token1Address]: {
+                  name: 'Token One',
+                  symbol: 'TOKEN1',
+                  decimals: 18,
+                  address: token1Address,
+                  iconUrl: 'https://token1.com/logo.png',
+                  occurrences: 1,
+                  aggregators: [],
+                },
                 [token2Address]: {
                   name: 'Hello Token',
                   symbol: 'HELLO',
@@ -258,10 +266,10 @@ describe('BridgeDestTokenSelector', () => {
     bridge: {
       sourceAmount: undefined,
       destAmount: undefined,
+      destChainId: undefined,
       sourceToken: undefined,
       destToken: undefined,
-      selectedSourceChainIds: [mockSourceChainId, mockDestChainId],
-      selectedDestChainId: mockDestChainId,
+      selectedSourceChainIds: undefined,
     },
   };
 
@@ -271,9 +279,9 @@ describe('BridgeDestTokenSelector', () => {
 
   it('renders with initial state and displays tokens', async () => {
     const { getByText, toJSON } = renderScreen(
-      BridgeDestTokenSelector,
+      BridgeSourceTokenSelector,
       {
-        name: Routes.BRIDGE.MODALS.DEST_TOKEN_SELECTOR,
+        name: Routes.BRIDGE.MODALS.SOURCE_TOKEN_SELECTOR,
       },
       { state: initialState }
     );
@@ -281,9 +289,19 @@ describe('BridgeDestTokenSelector', () => {
     // Header should be visible
     expect(getByText('Select token')).toBeTruthy();
 
-    // Tokens for destination chain should be visible
+    // Native token (ETH) should be visible with correct balance
     await waitFor(() => {
-      expect(getByText('HELLO')).toBeTruthy();
+      expect(getByText('3 ETH')).toBeTruthy();
+      expect(getByText('$6000')).toBeTruthy();
+    });
+
+    // ERC20 tokens should be visible with correct balances
+    await waitFor(() => {
+      expect(getByText('1 TOKEN1')).toBeTruthy();
+      expect(getByText('$20000')).toBeTruthy();
+
+      expect(getByText('2 HELLO')).toBeTruthy();
+      expect(getByText('$200000')).toBeTruthy();
     });
 
     expect(toJSON()).toMatchSnapshot();
@@ -291,79 +309,42 @@ describe('BridgeDestTokenSelector', () => {
 
   it('handles token selection correctly', async () => {
     const { getByText } = renderScreen(
-      BridgeDestTokenSelector,
+      BridgeSourceTokenSelector,
       {
-        name: Routes.BRIDGE.MODALS.DEST_TOKEN_SELECTOR,
+        name: Routes.BRIDGE.MODALS.SOURCE_TOKEN_SELECTOR,
       },
       { state: initialState }
     );
 
     await waitFor(() => {
-      const token1Element = getByText('HELLO');
+      const token1Element = getByText('1 TOKEN1');
       fireEvent.press(token1Element);
     });
 
-    expect(setDestToken).toHaveBeenCalledWith(expect.objectContaining({
-      address: token2Address,
-      aggregators: ['uniswap'],
-      balance: '2',
-      chainId: mockDestChainId,
+    expect(setSourceToken).toHaveBeenCalledWith({
+      address: token1Address,
+      aggregators: ['1inch'],
+      balance: '1',
+      balanceFiat: '$20000',
+      chainId: '0x1',
       decimals: 18,
-      image: 'https://token2.com/logo.png',
-      name: 'Hello Token',
-      symbol: 'HELLO',
-    }));
-    expect(mockGoBack).toHaveBeenCalled();
-  });
-
-  it('handles info button click correctly, navigates to Asset screen', async () => {
-    const { getAllByTestId, getByText } = renderScreen(
-      BridgeDestTokenSelector,
-      {
-        name: Routes.BRIDGE.MODALS.DEST_TOKEN_SELECTOR,
-      },
-      { state: initialState }
-    );
-
-    await waitFor(() => {
-      expect(getByText('ETH')).toBeTruthy();
-    });
-
-    // Get the info button using its test ID
-    const infoButton = getAllByTestId('token-info-button')[0];
-
-    // Ensure we found the info button
-    expect(infoButton).toBeTruthy();
-
-    // Press the info button
-    fireEvent.press(infoButton);
-
-    // Verify navigation to Asset screen with the correct token params
-    expect(mockNavigate).toHaveBeenCalledWith('Asset', expect.objectContaining({
-      address: '0x0000000000000000000000000000000000000000',
-      aggregators: [],
-      balance: '20',
-      balanceFiat: '$40000',
-      chainId: '0xa',
-      decimals: 18,
-      image: '',
-      isETH: true,
-      isNative: true,
+      image: 'https://token1.com/logo.png',
+      isETH: false,
+      isNative: false,
       isStaked: false,
-      logo: '../images/eth-logo-new.png',
-      name: 'Ethereum',
-      stakedBalance: '0x0',
-      symbol: 'ETH',
-      ticker: 'ETH',
-      tokenFiatAmount: 40000
-    }));
+      name: 'Token One',
+      symbol: 'TOKEN1',
+      token: 'Token One',
+      tokenFiatAmount: 20000,
+    });
+    expect(mockGoBack).toHaveBeenCalled();
   });
 
   it('handles close button correctly', () => {
     const { getByTestId } = renderScreen(
-      BridgeDestTokenSelector,
+      BridgeSourceTokenSelector,
       {
-        name: Routes.BRIDGE.MODALS.DEST_TOKEN_SELECTOR,
+        name: Routes.BRIDGE.MODALS.SOURCE_TOKEN_SELECTOR,
       },
       { state: initialState }
     );
@@ -376,41 +357,43 @@ describe('BridgeDestTokenSelector', () => {
 
   it('handles token search functionality correctly', async () => {
     const { getByTestId, getByText, queryByText } = renderScreen(
-      BridgeDestTokenSelector,
+      BridgeSourceTokenSelector,
       {
-        name: Routes.BRIDGE.MODALS.DEST_TOKEN_SELECTOR,
+        name: Routes.BRIDGE.MODALS.SOURCE_TOKEN_SELECTOR,
       },
       { state: initialState }
     );
 
     // Initially all tokens should be visible
     await waitFor(() => {
-      expect(getByText('HELLO')).toBeTruthy();
+      expect(getByText('3 ETH')).toBeTruthy();
+      expect(getByText('1 TOKEN1')).toBeTruthy();
+      expect(getByText('2 HELLO')).toBeTruthy();
     });
 
     // Search for TOKEN1
     const searchInput = getByTestId('bridge-token-search-input');
     fireEvent.changeText(searchInput, 'HELLO');
 
-    // Should only show HELLO, not ETH
+    // Should only show HELLO, not TOKEN1
     await waitFor(() => {
-      expect(getByText('HELLO')).toBeTruthy();
-      expect(queryByText('ETH')).toBeNull();
+      expect(getByText('2 HELLO')).toBeTruthy();
+      expect(queryByText('1 TOKEN1')).toBeNull();
     });
 
     // Search should be case-insensitive
     fireEvent.changeText(searchInput, 'hello');
     await waitFor(() => {
-      expect(getByText('HELLO')).toBeTruthy();
-      expect(queryByText('ETH')).toBeNull();
+      expect(getByText('2 HELLO')).toBeTruthy();
+      expect(queryByText('1 TOKEN1')).toBeNull();
     });
   });
 
   it('displays empty state when no tokens match search', async () => {
     const { getByTestId, getByText } = renderScreen(
-      BridgeDestTokenSelector,
+      BridgeSourceTokenSelector,
       {
-        name: Routes.BRIDGE.MODALS.DEST_TOKEN_SELECTOR,
+        name: Routes.BRIDGE.MODALS.SOURCE_TOKEN_SELECTOR,
       },
       { state: initialState }
     );
@@ -420,23 +403,6 @@ describe('BridgeDestTokenSelector', () => {
 
     await waitFor(() => {
       expect(getByText('No tokens match', { exact: false })).toBeTruthy();
-    });
-  });
-
-  it('navigates to destination network selector when See all is pressed', async () => {
-    const { getByText } = renderScreen(
-      BridgeDestTokenSelector,
-      {
-        name: Routes.BRIDGE.MODALS.DEST_TOKEN_SELECTOR,
-      },
-      { state: initialState }
-    );
-
-    const seeAllButton = getByText('See all');
-    fireEvent.press(seeAllButton);
-
-    expect(mockNavigate).toHaveBeenCalledWith(Routes.BRIDGE.MODALS.ROOT, {
-      screen: Routes.BRIDGE.MODALS.DEST_NETWORK_SELECTOR,
     });
   });
 });
