+++ app/components/UI/Bridge/components/BridgeTokenSelectorBase.tsx
@@ -1,19 +1,19 @@
 import React, { useCallback, useMemo, useRef } from 'react';
 import { StyleSheet, FlatList, TouchableOpacity } from 'react-native';
-import { Box } from '../Box/Box';
-import Text, { TextVariant, TextColor } from '../../../component-library/components/Texts/Text';
-import { useStyles } from '../../../component-library/hooks';
-import { Theme } from '../../../util/theme/models';
-import BottomSheetHeader from '../../../component-library/components/BottomSheets/BottomSheetHeader';
-import BottomSheet, { BottomSheetRef } from '../../../component-library/components/BottomSheets/BottomSheet';
-import { TokenI } from '../Tokens/types';
-import Icon, { IconName } from '../../../component-library/components/Icons/Icon';
-import { IconSize } from '../../../component-library/components/Icons/Icon/Icon.types';
-import { TokenIWithFiatAmount } from './useTokensWithBalance';
-import { strings } from '../../../../locales/i18n';
-import { FlexDirection, AlignItems, JustifyContent } from '../Box/box.types';
-import { useTokenSearch } from './useTokenSearch';
-import TextFieldSearch from '../../../component-library/components/Form/TextFieldSearch';
+import { Box } from '../../Box/Box';
+import Text, { TextVariant, TextColor } from '../../../../component-library/components/Texts/Text';
+import { useStyles } from '../../../../component-library/hooks';
+import { Theme } from '../../../../util/theme/models';
+import BottomSheetHeader from '../../../../component-library/components/BottomSheets/BottomSheetHeader';
+import BottomSheet, { BottomSheetRef } from '../../../../component-library/components/BottomSheets/BottomSheet';
+import { TokenI } from '../../Tokens/types';
+import Icon, { IconName } from '../../../../component-library/components/Icons/Icon';
+import { IconSize } from '../../../../component-library/components/Icons/Icon/Icon.types';
+import { TokenIWithFiatAmount } from '../hooks/useTokensWithBalance';
+import { strings } from '../../../../../locales/i18n';
+import { FlexDirection, AlignItems, JustifyContent } from '../../Box/box.types';
+import { useTokenSearch } from '../hooks/useTokenSearch';
+import TextFieldSearch from '../../../../component-library/components/Form/TextFieldSearch';
 
 const createStyles = (params: { theme: Theme }) => {
   const { theme } = params;
