+++ app/components/UI/Bridge/hooks/useGoToBridge.ts
@@ -7,9 +7,9 @@ import { useNavigation } from '@react-navigation/native';
 import { selectEvmChainId } from '../../../../selectors/networkController';
 
 import type { BrowserTab } from '../../Tokens/types';
-import type { BrowserParams } from '../../../../components/Views/Browser/Browser.types';
+import type { BrowserParams } from '../../../Views/Browser/Browser.types';
 import { getDecimalChainId } from '../../../../util/networks';
-import { useMetrics } from '../../../../components/hooks/useMetrics';
+import { useMetrics } from '../../../hooks/useMetrics';
 import { isBridgeUrl } from '../../../../util/url';
 
 /**
